#304: Moved and renamed the file containing ResistancesAlterations
It no longer belongs to the Objet area.
