+++ src/app/ffbe/mappers/consumable-mapper.ts
@@ -6,7 +6,7 @@ import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../ffbe.constan
 import {ItemCategoryFactory} from '../model/item-category.model';
 import {CategorieObjet} from '../model/objet/categorie-objet.model';
 import {FfbeUtils} from '../utils/ffbe-utils';
-import {ResistancesAlterations} from '../model/objet/objet-alterations-etat.model';
+import {ResistancesAlterations} from '../model/resistances-alterations.model';
 
 export class ConsumableMapper {
