+++ src/app/ffbe/mappers/equipment-mapper.ts
@@ -8,7 +8,7 @@ import {EquipmentStats} from '../model/equipment/equipment-stats.model';
 import {Caracteristiques} from '../model/caracteristiques.model';
 import {ResistancesElementaires} from '../model/resistances-elementaires.model';
 import {EquipmentElementResist} from '../model/equipment/equipment-element-resist.model';
-import {ResistancesAlterations} from '../model/objet/objet-alterations-etat.model';
+import {ResistancesAlterations} from '../model/resistances-alterations.model';
 import {EquipmentStatusEffect} from '../model/equipment/equipment-status-effect.model';
 import {Character} from '../model/character/character.model';
 import {CharactersService} from '../services/characters.service';
